pushing starts of process_results
